# docs/openapi.yaml
openapi: "3.0.1"
info:
  title: "API Documentation"
  version: "1.0"
paths:
...
  /click-counter:
    get:
      summary: "获取所有计数器"
      description: "获取所有计数器"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "array"
                    description: "响应数据"
                    items:
                      type: "object"
                      properties:
                        id:
                          type: "integer"
                          description: ""
                          format: "int64"
                        counterName:
                          type: "string"
                          description: ""
                        description:
                          type: "string"
                          description: ""
                        clickCount:
                          type: "integer"
                          description: ""
                          format: "int64"
                        enabled:
                          type: "boolean"
                          description: ""
                        lastClickTime:
                          type: "string"
                          description: ""
                        createdAt:
                          type: "string"
                          description: ""
                        updatedAt:
                          type: "string"
                          description: ""
                      description: "点击计数器响应DTO"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
    post:
      summary: "创建计数器"
      description: "创建计数器"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "counterName"
              type: "object"
              properties:
                counterName:
                  maxLength: 100
                  type: "string"
                  description: "计数器名称"
                description:
                  maxLength: 500
                  type: "string"
                  description: "计数器描述"
                enabled:
                  type: "boolean"
                  description: "是否启用"
                  default: "true"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: ""
                        format: "int64"
                      counterName:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      clickCount:
                        type: "integer"
                        description: ""
                        format: "int64"
                      enabled:
                        type: "boolean"
                        description: ""
                      lastClickTime:
                        type: "string"
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
  /click-counter/{id}:
    get:
      summary: "根据ID获取计数器"
      description: "根据ID获取计数器"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        schema:
          type: "integer"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: ""
                        format: "int64"
                      counterName:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      clickCount:
                        type: "integer"
                        description: ""
                        format: "int64"
                      enabled:
                        type: "boolean"
                        description: ""
                      lastClickTime:
                        type: "string"
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
    put:
      summary: "更新计数器"
      description: "更新计数器"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        schema:
          type: "integer"
      requestBody:
        content:
          application/json:
            schema:
              required:
              - "counterName"
              type: "object"
              properties:
                counterName:
                  maxLength: 100
                  type: "string"
                  description: "计数器名称"
                description:
                  maxLength: 500
                  type: "string"
                  description: "计数器描述"
                enabled:
                  type: "boolean"
                  description: "是否启用"
                  default: "true"
              description: ""
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: ""
                        format: "int64"
                      counterName:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      clickCount:
                        type: "integer"
                        description: ""
                        format: "int64"
                      enabled:
                        type: "boolean"
                        description: ""
                      lastClickTime:
                        type: "string"
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
    delete:
      summary: "删除计数器"
      description: "删除计数器"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        schema:
          type: "integer"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "string"
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
  /click-counter/name/{counterName}:
    get:
      summary: "根据名称获取计数器"
      description: "根据名称获取计数器"
      parameters:
      - name: "counterName"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: ""
                        format: "int64"
                      counterName:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      clickCount:
                        type: "integer"
                        description: ""
                        format: "int64"
                      enabled:
                        type: "boolean"
                        description: ""
                      lastClickTime:
                        type: "string"
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
...
  /click-counter/{id}/click:
    post:
      summary: "点击计数（根据ID）"
      description: "点击计数（根据ID）"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        schema:
          type: "integer"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: ""
                        format: "int64"
                      counterName:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      clickCount:
                        type: "integer"
                        description: ""
                        format: "int64"
                      enabled:
                        type: "boolean"
                        description: ""
                      lastClickTime:
                        type: "string"
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
  /click-counter/name/{counterName}/click:
    post:
      summary: "点击计数（根据名称）"
      description: "点击计数（根据名称）"
      parameters:
      - name: "counterName"
        in: "path"
        description: ""
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: ""
                        format: "int64"
                      counterName:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      clickCount:
                        type: "integer"
                        description: ""
                        format: "int64"
                      enabled:
                        type: "boolean"
                        description: ""
                      lastClickTime:
                        type: "string"
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
...
  /click-counter/page:
    get:
      summary: "分页查询计数器"
      description: "分页查询计数器"
      parameters:
      - name: "page"
        in: "query"
        description: ""
        required: true
        schema:
          type: "integer"
          nullable: false
          default: "1"
      - name: "size"
        in: "query"
        description: ""
        required: true
        schema:
          type: "integer"
          nullable: false
          default: "10"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "array"
                    description: "响应数据"
                    items:
                      type: "object"
                      properties:
                        id:
                          type: "integer"
                          description: ""
                          format: "int64"
                        counterName:
                          type: "string"
                          description: ""
                        description:
                          type: "string"
                          description: ""
                        clickCount:
                          type: "integer"
                          description: ""
                          format: "int64"
                        enabled:
                          type: "boolean"
                          description: ""
                        lastClickTime:
                          type: "string"
                          description: ""
                        createdAt:
                          type: "string"
                          description: ""
                        updatedAt:
                          type: "string"
                          description: ""
                      description: "点击计数器响应DTO"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
  /click-counter/search:
    get:
      summary: "根据条件查询计数器"
      description: "根据条件查询计数器"
      parameters:
      - name: "enabled"
        in: "query"
        description: ""
        required: false
        schema:
          type: "boolean"
          nullable: false
      - name: "counterName"
        in: "query"
        description: ""
        required: false
        schema:
          type: "string"
          nullable: false
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "array"
                    description: "响应数据"
                    items:
                      type: "object"
                      properties:
                        id:
                          type: "integer"
                          description: ""
                          format: "int64"
                        counterName:
                          type: "string"
                          description: ""
                        description:
                          type: "string"
                          description: ""
                        clickCount:
                          type: "integer"
                          description: ""
                          format: "int64"
                        enabled:
                          type: "boolean"
                          description: ""
                        lastClickTime:
                          type: "string"
                          description: ""
                        createdAt:
                          type: "string"
                          description: ""
                        updatedAt:
                          type: "string"
                          description: ""
                      description: "点击计数器响应DTO"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
  /click-counter/top:
    get:
      summary: "获取点击数最多的计数器"
      description: "获取点击数最多的计数器"
      parameters:
      - name: "limit"
        in: "query"
        description: ""
        required: true
        schema:
          type: "integer"
          nullable: false
          default: "10"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "array"
                    description: "响应数据"
                    items:
                      type: "object"
                      properties:
                        id:
                          type: "integer"
                          description: ""
                          format: "int64"
                        counterName:
                          type: "string"
                          description: ""
                        description:
                          type: "string"
                          description: ""
                        clickCount:
                          type: "integer"
                          description: ""
                          format: "int64"
                        enabled:
                          type: "boolean"
                          description: ""
                        lastClickTime:
                          type: "string"
                          description: ""
                        createdAt:
                          type: "string"
                          description: ""
                        updatedAt:
                          type: "string"
                          description: ""
                      description: "点击计数器响应DTO"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
  /click-counter/{id}/reset:
    post:
      summary: "重置计数器点击数"
      description: "重置计数器点击数"
      parameters:
      - name: "id"
        in: "path"
        description: ""
        required: true
        schema:
          type: "integer"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  code:
                    type: "integer"
                    description: "响应码"
                  message:
                    type: "string"
                    description: "响应消息"
                  data:
                    type: "object"
                    properties:
                      id:
                        type: "integer"
                        description: ""
                        format: "int64"
                      counterName:
                        type: "string"
                        description: ""
                      description:
                        type: "string"
                        description: ""
                      clickCount:
                        type: "integer"
                        description: ""
                        format: "int64"
                      enabled:
                        type: "boolean"
                        description: ""
                      lastClickTime:
                        type: "string"
                        description: ""
                      createdAt:
                        type: "string"
                        description: ""
                      updatedAt:
                        type: "string"
                        description: ""
                    description: "响应数据"
                  timestamp:
                    type: "integer"
                    description: "时间戳"
                    format: "int64"
...